Fixed errors within customers and tenants

+++ CoreAPI/Models/PointTransaction.cs
@@ -7,7 +7,7 @@ public record PointTransaction
     public string Id { get; private set; } = Guid.NewGuid().ToString();
     public string TenantId { get; private set; } = null!;
     public string CustomerId { get; private set; } = null!;
-    public decimal Amount { get; private set; }
+    public int Amount { get; private set; }
     public TransactionType Type { get; private set; }
     public string? Reason { get; private set; }
     public string? ReferenceId { get; private set; }
@@ -15,7 +15,7 @@ public record PointTransaction
 
     private PointTransaction() { }
 
-    public PointTransaction(string id, string tenantId, string customerId, decimal amount, TransactionType type, string? reason, string? referenceId = null)
+    public PointTransaction(string id, string tenantId, string customerId, int amount, TransactionType type, string? reason, string? referenceId = null)
     {
         Id = id;
         TenantId = tenantId;
@@ -26,6 +26,6 @@ public record PointTransaction
         ReferenceId = referenceId;
     }
     
-    public static PointTransaction Create(string tenantId, string customerId, decimal amount, TransactionType type, string? reason, string? referenceId = null)
+    public static PointTransaction Create(string tenantId, string customerId, int amount, TransactionType type, string? reason, string? referenceId = null)
         => new PointTransaction(Guid.NewGuid().ToString(), tenantId, customerId, amount, type, reason, referenceId);
 }
